+++ src/main/java/com/joshy/banking/dto/WithdrawRequest.java
@@ -1,15 +1,17 @@
 package com.joshy.banking.dto;
 
+import java.math.BigDecimal;
+
 public class WithdrawRequest {
     
     private String accountNumber;
-    private Double amount;
+    private BigDecimal amount;
 
     public String getAccountNumber() {
         return accountNumber;
     }
 
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
@@ -17,7 +19,7 @@ public class WithdrawRequest {
         this.accountNumber = accountNumber;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
 }
